+++ google-http-client/src/test/java/com/google/api/client/http/HttpResponseTest.java
@@ -29,6 +29,7 @@ import java.lang.reflect.Type;
 import java.nio.charset.StandardCharsets;
 import java.text.NumberFormat;
 import java.util.Arrays;
+import java.util.Locale;
 import java.util.logging.Level;
 import java.util.zip.GZIPInputStream;
 import java.util.zip.GZIPOutputStream;
@@ -57,6 +58,10 @@ public class HttpResponseTest extends TestCase {
 
   private static final String SAMPLE = "123\u05D9\u05e0\u05D9\u05D1";
   private static final String SAMPLE2 = "123abc";
+  private static final String VALID_CONTENT_TYPE = "text/plain";
+  private static final String VALID_CONTENT_TYPE_WITH_PARAMS =
+      "application/vnd.com.google.datastore.entity+json; charset=utf-8; version=v1; q=0.9";
+  private static final String INVALID_CONTENT_TYPE = "!!!invalid!!!";
 
   public void testParseAsString_utf8() throws Exception {
     HttpTransport transport =
@@ -99,6 +104,81 @@ public class HttpResponseTest extends TestCase {
         transport.createRequestFactory().buildGetRequest(HttpTesting.SIMPLE_GENERIC_URL);
     HttpResponse response = request.execute();
     assertEquals(SAMPLE2, response.parseAsString());
+  }
+
+  public void testParseAsString_validContentType() throws Exception {
+    HttpTransport transport =
+        new MockHttpTransport() {
+          @Override
+          public LowLevelHttpRequest buildRequest(String method, String url) throws IOException {
+            return new MockLowLevelHttpRequest() {
+              @Override
+              public LowLevelHttpResponse execute() throws IOException {
+                MockLowLevelHttpResponse result = new MockLowLevelHttpResponse();
+                result.setContent(SAMPLE2);
+                result.setContentType(VALID_CONTENT_TYPE);
+                return result;
+              }
+            };
+          }
+        };
+    HttpRequest request =
+        transport.createRequestFactory().buildGetRequest(HttpTesting.SIMPLE_GENERIC_URL);
+
+    HttpResponse response = request.execute();
+    assertEquals(SAMPLE2, response.parseAsString());
+    assertEquals(VALID_CONTENT_TYPE, response.getContentType());
+    assertNotNull(response.getMediaType());
+  }
+
+  public void testParseAsString_validContentTypeWithParams() throws Exception {
+    HttpTransport transport =
+        new MockHttpTransport() {
+          @Override
+          public LowLevelHttpRequest buildRequest(String method, String url) throws IOException {
+            return new MockLowLevelHttpRequest() {
+              @Override
+              public LowLevelHttpResponse execute() throws IOException {
+                MockLowLevelHttpResponse result = new MockLowLevelHttpResponse();
+                result.setContent(SAMPLE2);
+                result.setContentType(VALID_CONTENT_TYPE_WITH_PARAMS);
+                return result;
+              }
+            };
+          }
+        };
+    HttpRequest request =
+        transport.createRequestFactory().buildGetRequest(HttpTesting.SIMPLE_GENERIC_URL);
+
+    HttpResponse response = request.execute();
+    assertEquals(SAMPLE2, response.parseAsString());
+    assertEquals(VALID_CONTENT_TYPE_WITH_PARAMS, response.getContentType());
+    assertNotNull(response.getMediaType());
+  }
+
+  public void testParseAsString_invalidContentType() throws Exception {
+    HttpTransport transport =
+        new MockHttpTransport() {
+          @Override
+          public LowLevelHttpRequest buildRequest(String method, String url) throws IOException {
+            return new MockLowLevelHttpRequest() {
+              @Override
+              public LowLevelHttpResponse execute() throws IOException {
+                MockLowLevelHttpResponse result = new MockLowLevelHttpResponse();
+                result.setContent(SAMPLE2);
+                result.setContentType(INVALID_CONTENT_TYPE);
+                return result;
+              }
+            };
+          }
+        };
+    HttpRequest request =
+        transport.createRequestFactory().buildGetRequest(HttpTesting.SIMPLE_GENERIC_URL);
+
+    HttpResponse response = request.execute();
+    assertEquals(SAMPLE2, response.parseAsString());
+    assertEquals(INVALID_CONTENT_TYPE, response.getContentType());
+    assertNull(response.getMediaType());
   }
 
   public void testStatusCode_negative_dontThrowException() throws Exception {
@@ -461,17 +541,37 @@ public class HttpResponseTest extends TestCase {
   }
 
   public void testGetContent_gzipEncoding_finishReading() throws IOException {
+    do_testGetContent_gzipEncoding_finishReading("gzip");
+  }
+
+  public void testGetContent_gzipEncoding_finishReadingWithUppercaseContentEncoding() throws IOException {
+    do_testGetContent_gzipEncoding_finishReading("GZIP");
+  }
+
+  public void testGetContent_gzipEncoding_finishReadingWithDifferentDefaultLocaleAndUppercaseContentEncoding() throws IOException {
+    Locale originalDefaultLocale = Locale.getDefault();
+    try {
+      Locale.setDefault(Locale.forLanguageTag("tr-TR"));
+      do_testGetContent_gzipEncoding_finishReading("GZIP");
+    } finally {
+      Locale.setDefault(originalDefaultLocale);
+    }
+  }
+
+  private void do_testGetContent_gzipEncoding_finishReading(String contentEncoding) throws IOException {
     byte[] dataToCompress = "abcd".getBytes(StandardCharsets.UTF_8);
     byte[] mockBytes;
-    try (ByteArrayOutputStream byteStream = new ByteArrayOutputStream(dataToCompress.length)) {
-      GZIPOutputStream zipStream = new GZIPOutputStream((byteStream));
+    try (
+        ByteArrayOutputStream byteStream = new ByteArrayOutputStream(dataToCompress.length);
+        GZIPOutputStream zipStream = new GZIPOutputStream((byteStream))
+    ) {
       zipStream.write(dataToCompress);
       zipStream.close();
       mockBytes = byteStream.toByteArray();
     }
     final MockLowLevelHttpResponse mockResponse = new MockLowLevelHttpResponse();
     mockResponse.setContent(mockBytes);
-    mockResponse.setContentEncoding("gzip");
+    mockResponse.setContentEncoding(contentEncoding);
     mockResponse.setContentType("text/plain");
 
     HttpTransport transport =
@@ -490,9 +590,35 @@ public class HttpResponseTest extends TestCase {
     HttpRequest request =
         transport.createRequestFactory().buildHeadRequest(HttpTesting.SIMPLE_GENERIC_URL);
     HttpResponse response = request.execute();
-    TestableByteArrayInputStream output = (TestableByteArrayInputStream) mockResponse.getContent();
-    assertFalse(output.isClosed());
+    try (TestableByteArrayInputStream output = (TestableByteArrayInputStream) mockResponse.getContent()) {
+      assertFalse(output.isClosed());
+      assertEquals("abcd", response.parseAsString());
+      assertTrue(output.isClosed());
+    }
+  }
+
+  public void testGetContent_otherEncodingWithgzipInItsName_GzipIsNotUsed() throws IOException {
+    final MockLowLevelHttpResponse mockResponse = new MockLowLevelHttpResponse();
+    mockResponse.setContent("abcd");
+    mockResponse.setContentEncoding("otherEncodingWithgzipInItsName");
+    mockResponse.setContentType("text/plain");
+
+    HttpTransport transport =
+        new MockHttpTransport() {
+          @Override
+          public LowLevelHttpRequest buildRequest(String method, final String url)
+              throws IOException {
+            return new MockLowLevelHttpRequest() {
+              @Override
+              public LowLevelHttpResponse execute() throws IOException {
+                return mockResponse;
+              }
+            };
+          }
+        };
+    HttpRequest request = transport.createRequestFactory().buildHeadRequest(HttpTesting.SIMPLE_GENERIC_URL);
+    // If gzip was used on this response, an exception would be thrown
+    HttpResponse response = request.execute();
     assertEquals("abcd", response.parseAsString());
-    assertTrue(output.isClosed());
   }
 }