+++ google-http-client/src/test/java/com/google/api/client/util/DateTimeTest.java
@@ -28,12 +28,6 @@ public class DateTimeTest extends TestCase {
 
   private TimeZone originalTimeZone;
 
-  public DateTimeTest() {}
-
-  public DateTimeTest(String testName) {
-    super(testName);
-  }
-
   @Override
   protected void setUp() throws Exception {
     originalTimeZone = TimeZone.getDefault();
@@ -154,6 +148,11 @@ public class DateTimeTest extends TestCase {
         DateTime.parseRfc3339(
             "2018-12-31T23:59:59.9999Z"), // This value would be truncated prior to version 1.30.2
         DateTime.parseRfc3339("2018-12-31T23:59:59.999Z"));
+
+    // The beginning of Gregorian Calendar
+    assertEquals(
+        -12219287774877L, // Result from Joda time's Instant.parse
+        DateTime.parseRfc3339("1582-10-15T01:23:45.123Z").getValue());
   }
 
   /**
@@ -225,8 +224,18 @@ public class DateTimeTest extends TestCase {
     assertParsedRfc3339(
         "2018-03-01T10:11:12.1000Z", SecondsAndNanos.ofSecondsAndNanos(1519899072L, 100000000));
   }
+  
+  public void testEpoch() {
+    assertParsedRfc3339(
+        "1970-01-01T00:00:00.000Z", SecondsAndNanos.ofSecondsAndNanos(0, 0));
+  }
 
-  private void assertParsedRfc3339(String input, SecondsAndNanos expected) {
+  public void testOneSecondBeforeEpoch() {
+    assertParsedRfc3339(
+        "1969-12-31T23:59:59.000Z", SecondsAndNanos.ofSecondsAndNanos(-1, 0));
+  }
+
+  private static void assertParsedRfc3339(String input, SecondsAndNanos expected) {
     SecondsAndNanos actual = DateTime.parseRfc3339ToSecondsAndNanos(input);
     assertEquals(
         "Seconds for " + input + " do not match", expected.getSeconds(), actual.getSeconds());
@@ -249,7 +258,7 @@ public class DateTimeTest extends TestCase {
     assertEquals(expected, output);
   }
 
-  private void expectExceptionForParseRfc3339(String input) {
+  private static void expectExceptionForParseRfc3339(String input) {
     try {
       DateTime.parseRfc3339(input);
       fail("expected NumberFormatException");